cleaner separation between mock/actual impls
