+++ src/managers.rs
