+++ src/managers/emg.rs
